+++ pkg/analyze/dir_unix.go
@@ -14,7 +14,7 @@ func setPlatformSpecificAttrs(file *File, f os.FileInfo) {
 	switch stat := f.Sys().(type) {
 	case *syscall.Stat_t:
 		file.Usage = stat.Blocks * devBSize
-		file.Mtime = time.Unix(stat.Mtimespec.Sec, stat.Mtimespec.Nsec)
+		file.Mtime = time.Unix(int64(stat.Mtimespec.Sec), int64(stat.Mtimespec.Nsec))
 
 		if stat.Nlink > 1 {
 			file.Mli = stat.Ino
@@ -28,5 +28,5 @@ func setDirPlatformSpecificAttrs(dir *Dir, path string) {
 		return
 	}
 
-	dir.Mtime = time.Unix(stat.Mtimespec.Sec, stat.Mtimespec.Nsec)
+	dir.Mtime = time.Unix(int64(stat.Mtimespec.Sec), int64(stat.Mtimespec.Nsec))
 }
